tasks/requirements.txt
lxml
packaging
python-dateutil
remotezip
requests
python-libarchive